Add SpendingArgs to protocol.rs file
