+++ src/builder/mod.rs
@@ -1,5 +1,4 @@
 mod builder;
 mod protocol;
-mod spending_args;
 
-pub use self::{builder::ProtocolBuilder, spending_args::SpendingArgs};
+pub use self::{builder::ProtocolBuilder, protocol::SpendingArgs};
